third_party/WebKit/LayoutTests/animations/web-animations/element-animate-keyframe-value-warning.html

<div id="target"></div>
This test passes if there are three warnings about invalid keyframe values.
<script>
if (window.testRunner) {
  testRunner.dumpAsText();
}
target.animate({offsetPath: ''});
target.animate({offsetPath: 'path("m0 0")'});
target.animate({offsetPath: 'path("invalid")'});
target.animate({'--valid': ''});
target.animate({'--x': 'valid'});
target.animate({'--x': 'invalid )'});
target.animate({'--x': 'invalid !important'});
target.animate([{background: 'green', backgroundColor: 'green'}]);
</script>
