third_party/WebKit/LayoutTests/animations/web-animations/element-animate-keyframe-value-warning-expected.txt

CONSOLE WARNING: line 9: Invalid keyframe value for property offsetPath: path("invalid")
CONSOLE WARNING: line 12: Invalid keyframe value for property --x: invalid )
CONSOLE WARNING: line 13: Invalid keyframe value for property --x: invalid !important
This test passes if there are three warnings about invalid keyframe values.
